I think all the code is here for Level 0 - now I need to try it out to see if it still works! Then I need to finish/edit documentation because a lot of the original documentation is probably kaput by now.
NEXT: Level 1

In Level 1, we add in testing and more comprehensive documentation.

Every Player method should be unit-tested (with Minitest), including any edge cases. Make sure to write both tests that should fail and tests that should pass.
Every method in the Player class should adhere to SRP and DRY, and each method should only be able to return one type of data (for example: a method which can return one of three possible strings is good, but a method that might return a string, an integer, or a boolean depending on the input is a method that probably needs to be reconsidered).